+++ src/reporter.js
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2024 EPAM Systems
+ *  Copyright 2025 EPAM Systems
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -20,16 +20,12 @@ const getOptions = require('./utils/getOptions');
 const ReportingApi = require('./reportingApi');
 const {
   getAgentOptions,
-  getSuiteStartObject,
   getStartLaunchObject,
-  getTestStartObject,
-  getStepStartObject,
   getAgentInfo,
   getCodeRef,
-  getFullTestName,
   getFullStepName,
 } = require('./utils/objectUtils');
-const { TEST_ITEM_STATUSES, LOG_LEVEL } = require('./constants');
+const { TEST_ITEM_STATUSES, LOG_LEVEL, TEST_ITEM_TYPES } = require('./constants');
 
 const promiseErrorHandler = (promise) => {
   promise.catch((err) => {
@@ -43,8 +39,8 @@ class JestReportPortal {
     this.reportOptions = getAgentOptions(getOptions.options(options));
     this.client = new RPClient(this.reportOptions, agentInfo);
     this.tempSuiteIds = new Map();
-    this.tempTestIds = new Map();
     this.tempStepIds = new Map();
+    // TODO: Remove and use `this.tempStepIds` instead.
     this.tempStepId = null;
     this.promises = [];
 
@@ -60,70 +56,65 @@ class JestReportPortal {
     this.promises.push(promise);
   }
 
-  // Not called for `skipped` and `todo` specs
-  onTestCaseStart(test, testCaseStartInfo) {
-    if (testCaseStartInfo.ancestorTitles.length > 0) {
-      this._startSuite(testCaseStartInfo.ancestorTitles[0], test.path);
-    }
-    if (testCaseStartInfo.ancestorTitles.length > 1) {
-      this._startTest(testCaseStartInfo, test.path);
-    }
+  // FYI. In most cases it is not even called. It cannot be used for suites handling.
+  onTestStart() {}
 
-    const isRetried = !!this.tempStepIds.get(getFullStepName(testCaseStartInfo));
+  _startSuites(suiteTitles, filePath, startTime) {
+    suiteTitles.reduce((suitePath, suiteTitle) => {
+      const fullSuiteName = suitePath ? `${suitePath}/${suiteTitle}` : suiteTitle;
+      const codeRef = getCodeRef(filePath, fullSuiteName);
+      const parentCodeRef = getCodeRef(filePath, suitePath);
 
-    this._startStep(testCaseStartInfo, isRetried, test.path);
+      this._startSuite(suiteTitle, codeRef, parentCodeRef, startTime);
+      return fullSuiteName;
+    }, '');
   }
 
+  // TODO: cover with tests
+  // Not called for `skipped` and `todo` specs
+  onTestCaseStart(test, testCaseStartInfo) {
+    this._startSuites(testCaseStartInfo.ancestorTitles, test.path, testCaseStartInfo.startedAt);
+
+    const isRetried = !!this.tempStepIds.get(getFullStepName(testCaseStartInfo));
+    this._startStep(testCaseStartInfo, test.path, isRetried);
+  }
+
+  // TODO: cover with tests
   // Not called for `skipped` and `todo` specs
   onTestCaseResult(test, testCaseStartInfo) {
     this._finishStep(testCaseStartInfo);
   }
 
-  // Handling `skipped` tests and their ancestors
   onTestResult(test, testResult) {
-    let suiteDuration = 0;
-    let testDuration = 0;
-
+    // Handling `skipped` tests and their ancestors
     const skippedTests = testResult.testResults.filter(
       (t) => t.status === TEST_ITEM_STATUSES.SKIPPED,
     );
 
-    for (let index = 0; index < skippedTests.length; index++) {
-      const currentTest = skippedTests[index];
+    skippedTests.forEach((testCaseInfo) => {
+      const testCaseWithStartTime = { startedAt: new Date().valueOf(), ...testCaseInfo };
+      this._startSuites(testCaseInfo.ancestorTitles, test.path, testCaseWithStartTime.startedAt);
 
-      suiteDuration += currentTest.duration;
-      if (currentTest.ancestorTitles.length !== 1) {
-        testDuration += currentTest.duration;
-      }
-    }
+      if (testCaseWithStartTime.invocations) {
+        for (let i = 0; i < testCaseWithStartTime.invocations; i++) {
+          const isRetried = i > 0;
 
-    skippedTests.forEach((t) => {
-      if (t.ancestorTitles.length > 0) {
-        this._startSuite(t.ancestorTitles[0], test.path, suiteDuration);
-      }
-      if (t.ancestorTitles.length > 1) {
-        this._startTest(t, test.path, testDuration);
-      }
-
-      if (!t.invocations) {
-        this._startStep(t, false, test.path);
-        this._finishStep(t);
-        return;
-      }
-
-      for (let i = 0; i < t.invocations; i++) {
-        const isRetried = t.invocations !== 1;
-
-        this._startStep(t, isRetried, test.path);
-        this._finishStep(t);
+          this._startStep(testCaseWithStartTime, test.path, isRetried);
+          this._finishStep(testCaseWithStartTime);
+        }
+      } else {
+        this._startStep(testCaseWithStartTime, test.path);
+        this._finishStep(testCaseWithStartTime);
       }
     });
 
-    this.tempTestIds.forEach((tempTestId, key) => {
-      this._finishTest(tempTestId, key);
-    });
-    this.tempSuiteIds.forEach((tempSuiteId, key) => {
-      this._finishSuite(tempSuiteId, key);
+    const suiteFilePathToFinish = getCodeRef(testResult.testFilePath);
+
+    // Finishing suites that are related to the test file
+    this.tempSuiteIds.forEach((suiteTempId, suiteFullName) => {
+      if (suiteFullName.includes(suiteFilePathToFinish)) {
+        this._finishSuite(suiteTempId, suiteFullName);
+      }
     });
   }
 
@@ -138,54 +129,44 @@ class JestReportPortal {
     await promise;
   }
 
-  _startSuite(suiteName, path, suiteDuration) {
-    if (this.tempSuiteIds.get(suiteName)) {
-      return;
-    }
-    const codeRef = getCodeRef(path, suiteName);
-    const { tempId, promise } = this.client.startTestItem(
-      getSuiteStartObject(suiteName, codeRef, suiteDuration),
-      this.tempLaunchId,
-    );
-
-    this.tempSuiteIds.set(suiteName, tempId);
-    promiseErrorHandler(promise);
-    this.promises.push(promise);
-  }
-
-  _startTest(test, testPath, testDuration) {
-    if (this.tempTestIds.get(test.ancestorTitles.join('/'))) {
+  _startSuite(title, codeRef, parentCodeRef = '', startTime = new Date().valueOf()) {
+    if (this.tempSuiteIds.get(codeRef)) {
       return;
     }
 
-    const tempSuiteId = this.tempSuiteIds.get(test.ancestorTitles[0]);
-    const fullTestName = getFullTestName(test);
-    const codeRef = getCodeRef(testPath, fullTestName);
-    const testStartObj = getTestStartObject(
-      test.ancestorTitles[test.ancestorTitles.length - 1],
+    const testStartObj = {
+      type: TEST_ITEM_TYPES.SUITE,
+      name: title,
       codeRef,
-      testDuration,
-    );
-    const parentId =
-      this.tempTestIds.get(test.ancestorTitles.slice(0, -1).join('/')) || tempSuiteId;
+      startTime,
+    };
+    const parentId = this.tempSuiteIds.get(parentCodeRef);
     const { tempId, promise } = this.client.startTestItem(
       testStartObj,
       this.tempLaunchId,
       parentId,
     );
 
-    this.tempTestIds.set(fullTestName, tempId);
+    this.tempSuiteIds.set(codeRef, tempId);
     promiseErrorHandler(promise);
     this.promises.push(promise);
   }
 
-  _startStep(test, isRetried, testPath) {
-    const tempSuiteId = this.tempSuiteIds.get(test.ancestorTitles[0]);
+  _startStep(test, testPath, isRetried = false) {
     const fullStepName = getFullStepName(test);
     const codeRef = getCodeRef(testPath, fullStepName);
-    const stepDuration = test.duration;
-    const stepStartObj = getStepStartObject(test.title, isRetried, codeRef, stepDuration);
-    const parentId = this.tempTestIds.get(test.ancestorTitles.join('/')) || tempSuiteId;
+    const stepStartObj = {
+      type: TEST_ITEM_TYPES.STEP,
+      name: test.title,
+      codeRef,
+      startTime: test.startedAt,
+      retry: isRetried,
+    };
+
+    const parentFullName = test.ancestorTitles.join('/');
+    const parentCodeRef = getCodeRef(testPath, parentFullName);
+    const parentId = this.tempSuiteIds.get(parentCodeRef);
+
     const { tempId, promise } = this.client.startTestItem(
       stepStartObj,
       this.tempLaunchId,
@@ -237,7 +218,7 @@ class JestReportPortal {
         : `\`\`\`error\n${stripAnsi(failureMessage)}\n\`\`\``;
     const finishTestObj = { status, ...(description && { description }) };
 
-    this.sendLog({ message: failureMessage, level: LOG_LEVEL.ERROR, tempStepId });
+    this._sendLog({ message: failureMessage, level: LOG_LEVEL.ERROR, tempStepId });
 
     const { promise } = this.client.finishTestItem(tempStepId, finishTestObj);
 
@@ -245,7 +226,7 @@ class JestReportPortal {
     this.promises.push(promise);
   }
 
-  sendLog({ level = LOG_LEVEL.INFO, message = '', file, time, tempStepId }) {
+  _sendLog({ level = LOG_LEVEL.INFO, message = '', file, time, tempStepId }) {
     const newMessage = stripAnsi(message);
     const { promise } = this.client.sendLog(
       tempStepId === undefined ? this.tempStepId : tempStepId,
@@ -274,20 +255,12 @@ class JestReportPortal {
     this.promises.push(promise);
   }
 
-  _finishTest(tempTestId, key) {
-    if (!tempTestId) return;
+  _finishSuite(tempTestId, key) {
+    if (!tempTestId) {
+      return;
+    }
 
     const { promise } = this.client.finishTestItem(tempTestId, {});
-
-    this.tempTestIds.delete(key);
-    promiseErrorHandler(promise);
-    this.promises.push(promise);
-  }
-
-  _finishSuite(tempSuiteId, key) {
-    if (!tempSuiteId) return;
-
-    const { promise } = this.client.finishTestItem(tempSuiteId, {});
 
     this.tempSuiteIds.delete(key);
     promiseErrorHandler(promise);